refactor(cli): Clarify role of checks
