+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod checks;
 pub mod config;
 pub mod dict;
+pub mod file;
 pub mod report;
